implement very basic program list
In order to make the versioncontrol classes accessable and utils, all files and directories of build/src (excluding server and client) are now symlinked to the client directory. Moreover, defaultJSExtensions has been set to true in systemjs config.

This is only a temporary solution for accessing modules outside the client folder and will (hopefully) be updated to a cleaner alternative soon.